requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.0
pandas-datareader>=0.10.0
requests>=2.31.0
python-dotenv==1.0.1
pytz>=2024.1
aiohttp>=3.9.0
asyncio>=3.4.3
tqdm>=4.66.0
pyarrow>=15.0.0
ta-lib>=0.4.0
scikit-learn>=1.3.0
pyyaml>=6.0.0
joblib>=1.3.0
xgboost>=2.0.0
tensorflow>=2.15.0
keras>=2.15.0
pytest>=7.0.0
ta  # for technical analysis
